Make sure we have a file when starting, even if we didn't specify one
The problem solved here is that when starting the emacs daemon, some
settings are not applied (because no display is created for a file).

This means that some settings only run on server-visit-hook, so as long
as we visit a file, things go ok.

So, when we don't specify a file on the commandline, we make one up so
that hook runs and our settings are applied.