Revert "Beginning of Cask based configuration" and some
This reverts commit 55cb858ec1bffc865e77d015b5950fec1509839c.

Conflicts:
	Cask
	mrb.org

I tried using Cask, but I fail to see what's great about it. It
disconnects package management more than it needs to from emacs. I
haven't figured out a way to do the equivalent of el-get-install
interactively with Cask, so it's out the door and I'm back to el-get

Of course my git DAG was not properly set up to revert this cleanly, so
I had to correct a whole bunch of things by hand.