ws-butler can not be enabled globally
It encounters read only buffers for which the highlight changes mode can
not be enabled.
We should either create hooks per mode or submit a patch for the mode
which just looks at the read-only state and then disables ws-butler mode
for that buffer. For now, just not enabling the global mode.