Write functions for inserting a modeline
When emacs does not determine the right mode when visting a buffer;
happens regularly when I write a shell script that doesn't have the .sh
extension for example. I sometimes want a modestring somewhere in the
file. Typically that string gets inserted on a line which is commented
out in the proper syntax for that mode.

mrb/insert-mode-string does that