Change ido default buffer methode to ask to jump
This covers a situation where there are multiple emacs frames and I
want to select a buffer in multiple frames for some reason.

Now, emacs will ask me wheter to jump to the other frame or open the
buffer in the current one. I typically want it to open in the current
one.