Activate evil support in which-key package
This is perhaps even more important thant the normal completion for
emacs prefix keys, which I know a little bit better.

The delay of 2 seconds should get longer (or shorter?) once I get more
familiar with the keybindings. Waiting for the menu is now a conscious
decision, so it's long enough.