Show 1 attention seeking buffer at a time in full
The tracking buffer list is a stack so we won't lose information, but
I only need to see one at a time, unabbreviated.

If there are more buffers needing attention there will be a counter
next to the first.

Only usage at this point are weechat buffers, but the mechanism is
generic and user controllable, so I guess I'll use this for other
things over time as well.

One thing that popped in my mind was having a notifier if the contents
of a buffer changed (like a reverted to disk version)
Another is 'twittering' events (like mentioning)