Performance experiment
This makes everything seem very fast, but is probably going to lead to
problems.

What it does:
- make the gc threshold very large (500Mb)
- set an idle timer to garbage collect when emacs is idle

So, when emacs is doing stuff it basically does not garbage collect
untill we have 500Mb of the stuff (which is only during longer streaks
of code execution I guess).

Once emacs is idle for a second or 5, we spent another second cleaning
up stuff.

This solves most of my cursor stuttering problems and makes many
operations in orgmode a lot faster.