[notmuch] Refresh buffer after tag changes
Not sure if this is a keeper. It resolves having to guess if a message
is deleted and or spam (because it will disappear from the list), but
probably will annoy me on larger sets (performance)

I think this will still work on having a range selected and then using
the shortcut, but need to test this.