Make sure we update the cache on startup
As we are not updating the smex cache automatically anymore, make sure
it gets updated at least once on startup, so we start out with a full
set of available commands.

I'm not sure if this is needed, it should be automatic, but I think it
is not. Oh well.