Adjust automatically carryover items
Added a match for a 'CARRYOVER' property so I can mark headlines to
carry over to the next day in the journal.

This as such is useful, but it got triggered by something else I was
doing.

On capturing new items, they end up in a header `INBOX` in may main
org file, from where they should end up being executed if they are
small things, or be refiled to somewhere else to pop up in later
reviews or be made in to a project if the task is larger.

I found that the INBOX method above wasn't very helpful in daily
chores type of things. What I usually do is, with my first cup of
coffee, make a list of things I want to do today. I do this on paper,
an I want to get rid of that. This list often contains something like:
"look in org for things" or "orgmode organisation"

So, here's what I came up with to improve this and get rid of the
paper thing.

First, I create an org-journal entry under which I keep that list of
daily things, typically 5-10 items. To allow adding things to that
list I put an ID of `new-todo-receiver` on the heading. That ID is
used by my capture template for new TODO items. So, effectively I have
now moved my INBOX to org-journal and can check it the whole day while
making notes about them in the same file as journal entries.

Come the next day, I obviously want the same thing. So, I need to
carry over the whole list. By default org-journal carries TODO items
over to the next day, so I could have it working just by making the
header, which has the capture ID, a 'TODO' item.

This change allows me to omit that TODO on the header, which looks a
bit nicer, but it also allows 'work in progress' text sections to move
with me every day without having them show up in TODO queries and
without me having to look them up every time. To do this I just put a
PROPERTY on the relevant heading of `CARRYOVER="yes"` and org-journal
will carry it over the next day. Once the text if finished it gets
published, emailed or refiled.