Dramatic speedup for finding my capture location
My ID is in the latest org-journal file, so that heuristic is used to
find the proper capture location a lot faster than scanning all files
which the (id "id-value") presumably does.

By doing a org-journal-new-entry without creating a journal entry, I
get the effect that the file is directly there, and I just have to
find the ID there to be able to capture.