Centralize cursor type and use a bar instead of box
Using a box is sometimes confusing as you are led to believe the point
is *on* a character whereas it actually is *between* two characters.

This can get especially confusing when using paren matching
packages. The terms 'inside' and 'outside' are just more clear when it
is apparent where the point actually is.

I've used a bar with a width of 5 (pixels?) to make it a bit more
visible. The default bar is very thin.