Archive evil configuration for now
I have tried it for 6 months or so and the added value for me is
negligible. I wanted it to work, but the mode keep getting in the way
for me.

I have basically the same issue, for example, in qutebrowser, where
I'm often typing in input fields during normal mode. Very frustrating.

Next to that nuisance, I have quite a couple of emacs packages which
have their own state. Like:
- magit
- magithub
- notmuch
- elfeed

These packages do not benefit at all from evil, without major
reconfiguration.