Add function which for org-capture-templates to insert link info
Usecase:

I want to insert the /source/ of the TODO item with a link back to
mu4e for example, but insert nothing if there is no source.

Evaluation within org-capture templates is a bit tricky, so a helper
function is needed which extract info from the plist which holds the
link info.