Add custom mark to mark message as Spam
In the current config, this means that the message is moved to the
"/Junk" folder. When the mailstore is synced to the IMAP server, the
setup there will learn that the message should have been marked as
spam.

Marking as Ham is moving message *out* of the "/Junk" folder which can
just be done with the 'move' mark. The destionation can then be chosen
as whatever folder is fitting for that message.