Beginnings of LSP usage
The Language Server Protocol in Emacs seems like a good way to deal
with context information, reference and syntax on languages in a
consistent way. Starting out with some random buffers I had open at
the time, but eventually I want to add this for every language I use.

The elephant in the room here is language servers NOT being
implemented for emacs lisp itself and other lisp like languages (I've
found one for racket though)