Make tags column dynamic, most of the time
Align tags according to `window-body-width`, which also works in
visual line / column mode. This effectively makes the tags right
aligned and conforming to the visual wrapping I do.

It's not enough though, the focus events are quite arbitrarily chosen,
it should probably on some idle timer, but now I can trigger is when I
want for now.