[mu4e] Hide personal addrs and use current context on compose
The policy was 'pick-first (when nothing matches) and is now 'use current' (still when nothing
matches). This allows me to choose a context and start composing.

The 'default' context is still used for 95% of my usecases.

This should not change the configured behaviour when a context *does* match, i.e. in the case that
an explicitly addressed email is replied to.