Make aligning tags by hand a bit easier
The automatic align I had in place earlier did not work out, so back to manual.

I use this quite often, annoyingly (because it messes up git commit logs with spacing) so I probably
want a binding for this later on, but I'm not sure yet which one to use for it

S-M q aligns the paragraph, so something in that neighborhood I guess.