Use async native compilation, when available
I'm experimenting with the native-comp branch which promises to speed things up quite a bit. This
setting enables asynchronous automatic compilation of .el files to their native equivalent and
replaces the elisp forms for their native ones when done.

Compilation needs to happen at least once which takes a while on first startup (asynchronously, so
it's almost invisible). After finishing the compilation, emacs should automatically load the .eln
files which are the "binaries"

I have no good way to measure the performance difference in real life, but the artificial benchmarks
are very promising.

In theory this should be the only change needed, but I can see a need to have a notification when
compilation happens (like in the modeline or something like that)