Archiving eshell configuration
I have never used it much and I'm more used to bash and zsh work. With the config of vterm I think
all my terminal and shell needs are met.

The config was archived in /the org way/ so it stays availbe in my org archives should I want to
revive it at some point.