Explicitly set the outline colors
I am staring at these all day, so some care must be taken on customizing these. To my surprise,
level 8 is actually used quite a bit.

I try to use the redder tones to lower levels, signifying that I should flatten the tree a
bit (reddish does not work well for headings).

The blues and greens are more prevalent in the lower levels which are used most often