Allow specific keybindings for active regions
The way this is organized now, as in all with the selected package itself is not how I would like to
do it. I'd rather have the mode specific bindings with the specific package, but I could not get
that to work, so I'll have to look a that later.