Do the last updated timestamp smarter
Got annoyed with the constant noise of it in diffs.

Now it uses a named code block in orgmode which gets the last commit date and inserts that on
exporting the file.

The downside is that I need to confirm executing this code on export. Maybe I'll wrap the export for
the file in a function that disables the confirmation temporarily.