Replace paredit with puni+electric-pair mode
Got a bit fed up with all the exceptions in other modes needed for paredit

Puni seemed a cleaner solution, also not assuming very much about the mode, so this will work
regardless of the language.

The autopairing is handled by the built-in electric-pair-mode now

With this I get the same features as with paredit (give or take a few keybindings that need to be
made along the way).