Reconfigure mode-line and tab-bar
Using modeline for buffer specific info only, tab-bar holds global and possibly mode-specific
info. This resolves a long time itch of duplicating status information in all of the modelines on
screen.

Using nano-modeline for the modeline and built-in tab-bar for the statusbar

There's many small things that can now be simpler because of this.