Change obsolete variables
Trying the new buffer type switch for now, which is just focussing the new buffer window whereas
'window is not focussing the new buffer.

new frame is too much as on send it does not get closed, so I end up with many frames.

when the latter is fixed/changed I think I would prefer frame

mu4e-compose-signature is not needed at all anymore, message-signature has it